get old py scripts working with fspython
