+++ fsfast/bin/par2schedule
@@ -1,4 +1,4 @@
-#! /usr/bin/python
+#!/usr/bin/env python
 
 #
 # par2schedule
