+++ scripts/CMakeLists.txt
@@ -2,11 +2,9 @@ install(PROGRAMS
   annot2std
   aparc2feat
   aparc_stats_aseg
-  aparcstats2table
   aparcstatsdiff
   apas2aseg
   aseg2feat
-  asegstats2table
   asegstatsdiff
   avi2talxfm
   bblabel
@@ -29,7 +27,6 @@ install(PROGRAMS
   conf2hires
   cor_to_minc
   cp-dicom
-  datastruct_utils.py
   dcmdir-info-mgh
   dcmsplit
   dcmunpack
@@ -41,7 +38,6 @@ install(PROGRAMS
   dmrirc.long.example
   dsh
   dt_recon
-  easyfv
   epidewarp.fsl
   extract_talairach_avi_QA.awk
   feat2segstats
@@ -68,17 +64,14 @@ install(PROGRAMS
   fscalc.fsl
   fsdcmdecompress
   fsfget
-  fsgd_parser.py
   fsl.5.0.2.xyztrans.sch
   fsl_rigid_register
   fsl_sub_mgh
   fslregister
-  fspalm
   fsr-checkxopts
   fsr-coreg  
   fsr-getxopts
   fsr-import
-  fsutils.py
   fsvglrun
   fvcompare
   gca-apply
@@ -120,15 +113,7 @@ install(PROGRAMS
   listsubj
   long_create_base_sigma
   long_create_orig
-  long_mris_slopes
-  long_qdec_table
-  long_stats_combine
-  long_stats_slopes
-  long_stats_tps
-  long_submit_jobs
-  long_submit_postproc
   longmc
-  LongQdecTable.py
   lpcregister
   make-segvol-table
   make_average_subcort
@@ -146,10 +131,8 @@ install(PROGRAMS
   map_to_base
   mcparams2reg
   meanval
-  merge_stats_tables
   mergeseg
   minc2seqinfo
-  misc.py
   mkheadsurf
   mkima_index.tcl
   mkmnc_index.tcl
@@ -217,7 +200,6 @@ install(PROGRAMS
   renormalize_subject_keep_editting
   renormalize_T1_subject
   reregister_subject_mixed
-  rmcr
   rtview
   run-qdec-glm
   run_mris_preproc
@@ -245,7 +227,6 @@ install(PROGRAMS
   sratio
   stattablediff
   stem2fname
-  subject_info.py
   surfreg
   tal_compare
   talairach
@@ -261,7 +242,6 @@ install(PROGRAMS
   trac-all
   trac-paths
   trac-preproc
-  tractstats2table
   train-gcs-atlas
   unpack_ima.tcl
   unpack_ima1.tcl
@@ -291,7 +271,22 @@ install(PROGRAMS
   DESTINATION bin
 )
 
-install_pyscript(rca-config rca-config2csh)
+install_pyscript(
+  long_submit_jobs
+  long_qdec_table
+  long_submit_postproc
+  long_stats_tps
+  long_stats_slopes
+  long_mris_slopes
+  long_stats_combine
+  aparcstats2table
+  asegstats2table
+  tractstats2table
+  merge_stats_tables
+  fspalm
+  rca-config
+  rca-config2csh
+)
 
 install_append_help(bbregister bbregister.help.xml bin)
