+++ scripts/asegstats2table
@@ -1,16 +1,16 @@
-#!/usr/bin/env python2
-from __future__ import print_function
+#!/usr/bin/env python
+
 import warnings
 warnings.filterwarnings('ignore', '.*negative int.*')
+
 import os
 import sys
 import optparse
 import logging
 import csv
 
-# sys.path.append(os.environ['FREESURFER_HOME']+'/bin/')
-import fsutils
-from fsutils import AsegStatsParser, BadFileError, aseglogger
+import freesurfer.utils.legacy as fsutils
+from freesurfer.utils.legacy import AsegStatsParser, BadFileError, aseglogger
 
 # Original Version - Douglas Greve, MGH
 # Rewrite - Krish Subramaniam, MGH
