+++ scripts/long_stats_combine
@@ -1,5 +1,4 @@
-#!/usr/bin/env python2
-# -*- coding: latin-1 -*-
+#!/usr/bin/env python
 
 #
 # long_stats_tps
@@ -21,11 +20,12 @@
 # https://surfer.nmr.mgh.harvard.edu/fswiki/FreeSurferSoftwareLicense
 #
 # Reporting: freesurfer@nmr.mgh.harvard.edu
-#
-#
+
+
 from __future__ import print_function
 import warnings
 warnings.filterwarnings('ignore', '.*negative int.*')
+
 import os
 import sys
 import shlex
@@ -34,8 +34,9 @@ import logging
 import subprocess
 import tempfile
 import shutil
-from subject_info import *
-from LongQdecTable import *
+
+from freesurfer.utils.legacy import LongQdecTable
+
 
 # logging 
 ch = logging.StreamHandler()
