+++ scripts/long_submit_jobs
@@ -1,5 +1,4 @@
 #!/usr/bin/env python
-# -*- coding: latin-1 -*-
 
 #
 # long_submit_jobs
@@ -22,10 +21,11 @@
 #
 # Reporting: freesurfer@nmr.mgh.harvard.edu
 #
-#
+
 from __future__ import print_function
 import warnings
 warnings.filterwarnings('ignore', '.*negative int.*')
+
 import os
 import sys
 import shlex
@@ -38,11 +38,8 @@ import time
 import math
 import stat
 
-#sys.path.append(os.environ['FREESURFER_HOME']+'/bin/') 
-#sys.path.append('/usr/local/freesurfer/dev/bin/') 
+from freesurfer.utils.legacy import LongQdecTable
 
-from subject_info import *
-from LongQdecTable import *
 
 # logging 
 ch = logging.StreamHandler()
