+++ scripts/long_submit_postproc
@@ -1,5 +1,4 @@
-#!/usr/bin/env python2
-# -*- coding: latin-1 -*-
+#!/usr/bin/env python
 
 #
 # long_submit_stuff
@@ -22,10 +21,11 @@
 #
 # Reporting: freesurfer@nmr.mgh.harvard.edu
 #
-#
+
 from __future__ import print_function
 import warnings
 warnings.filterwarnings('ignore', '.*negative int.*')
+
 import os
 import sys
 import shlex
@@ -36,8 +36,9 @@ import tempfile
 import shutil
 import time
 import stat
-from subject_info import *
-from LongQdecTable import *
+
+from freesurfer.utils.legacy import LongQdecTable
+
 
 # logging 
 ch = logging.StreamHandler()
