--- scripts/rmcr
@@ -1,8 +1,0 @@
-#!/usr/bin/python
-# This script replaces carriage returns with line feeds
-# Usage: rmcr inputfile > outputfile > 
-# $Id: rmcr,v 1.1 2015/03/24 21:56:02 greve Exp $
-import sys; 
-import fileinput; 
-sys.stdout.writelines(line.replace('\r', '\n') for line in fileinput.input(mode='rU'))
-
