+++ scripts/stattablediff
@@ -1,4 +1,5 @@
-#!/usr/bin/python
+#!/usr/bin/env python
+
 import sys;
 import os;
 import string;
